Upload CoPy files
These files comprise a rudimentary game ("Champions of Pythonia," or "CoPy") that was written with Python in June of 2022. 
The objective was to leverage the various GUI capabilities of the tkinter module to create an aesthetically pleasing mini-RPG.
This program hasn’t been scrutinized for a year.  It may be revised, as time permits.